fix: move all test positions into header file
