+++ CMakeLists.txt
@@ -36,7 +36,6 @@ target_include_directories(vividmind PRIVATE
 
 add_executable(tests
     tests/test_move_gen.cpp
-    tests/test_positions.cpp
     ${COMMON_SOURCES}
 )
 target_include_directories(tests PRIVATE
